Remove courier references and add default accounts
Removing courier references as they are redundant as per the requirements
in the PDF in the repository. Now OWNER can list all users and tasks and
is the only account type which can register offices. CUSTOMER type users
will request from EMPLOYEE type user task for delivery. And the EMPLOYEE
should accept or decline the delivery task.

Default accounts are now added on startup so we escape the need for
registration forms if we want to update the logic with tasks creation.

Signed-off-by: Martin Dekov <mvdekov@gmail.com>